Export button
- ticket: https://reevoo.atlassian.net/browse/PROD-4910
- this commit adds export button above the tableau dashboards and allow users to export theirs data as csv, pdf or image
- it also removes the old unused functionality for creating and saving views inside tableau
- export button is at this moment in testing phase, that is why we enable this functionality only to those users having 'exportButton' key in theirs cookies
