+++ app/js/components/dashboard_exports/dashboard_exports.scss
@@ -1,4 +1,4 @@
-.dashboard-filters {
+.dashboard-exports {
   position: relative;
 
   &__overlay {
