+++ app/js/services/auth.js
@@ -19,6 +19,9 @@ const setAccessToken = (newToken) => Cookies.set('accessToken', newToken)
 
 const getRefreshToken = () => Cookies.get('refreshToken')
 
+// Until we have configuration options implemented, we enable export button only to test users
+export const exportEnabled = () => Cookies.get('exportEnabled') === 'true'
+
 /* The access token is a JWT token, which is three strings separated by ".":
  *   <jwt_metadata>.<payload>.<signature>
  *
